agent/requirements.txt
speedtest-cli
ping3
dnspython
schedule